+++ tests/test_checker/test_noqa.py
@@ -208,6 +208,7 @@ SHOULD_BE_RAISED = types.MappingProxyType({
     'WPS524': 1,
     'WPS525': 2,
     'WPS526': 1,
+    'WPS527': 1,
 
     'WPS600': 1,
     'WPS601': 1,