+++ CHANGELOG.md
@@ -30,6 +30,7 @@ We used to have incremental versioning before `0.1.0`.
 - Forbids to start lines with `.`
 - Enforces better `&`, `|`, `>>`, `<<`, `^` operators usage
 - Forbids incorrect exception order
+- Enforces tuples usage with frozenset constructor
 
 ### Bugfixes